move sharable sort parts
