+++ webui/react/src/components/HpSelectFilter.tsx
@@ -5,7 +5,7 @@ import React, { useCallback, useMemo } from 'react';
 import HumanReadableNumber from 'components/HumanReadableNumber';
 import { clone, isObject } from 'shared/utils/data';
 import { ALL_VALUE, HpImportance } from 'types';
-import { hpImportanceSorter } from 'utils/sort';
+import { hpImportanceSorter } from '../utils/sort';
 
 import css from './HpSelectFilter.module.scss';
 import SelectFilter, { Props as SelectFilterProps } from './SelectFilter';
