+++ webui/react/src/components/LogViewer/LogViewer.tsx
@@ -21,7 +21,7 @@ import { clone } from 'shared/utils/data';
 import { formatDatetime } from 'shared/utils/datetime';
 import { copyToClipboard } from 'shared/utils/dom';
 import { Log, LogLevel } from 'types';
-import { dateTimeStringSorter, numericSorter } from 'utils/sort';
+import { dateTimeStringSorter, numericSorter } from 'shared/utils/sort';
 
 import css from './LogViewer.module.scss';
 import LogViewerEntry, { DATETIME_FORMAT, ICON_WIDTH, MAX_DATETIME_LENGTH } from './LogViewerEntry';
