+++ webui/react/src/components/Table.stories.tsx
@@ -4,9 +4,10 @@ import React from 'react';
 
 import { useStore } from 'contexts/Store';
 import { CommandTask } from 'types';
-import { alphaNumericSorter, commandStateSorter, dateTimeStringSorter } from 'utils/sort';
+import { alphaNumericSorter, dateTimeStringSorter } from 'shared/utils/sort';
 import { generateCommandTask } from 'utils/task';
 import { getDisplayName } from 'utils/user';
+import { commandStateSorter } from '../utils/sort';
 
 import {
   defaultRowClassName, relativeTimeRenderer, stateRenderer,
