+++ webui/react/src/components/TagList.tsx
@@ -4,7 +4,7 @@ import React, { useCallback, useEffect, useRef, useState } from 'react';
 
 import Link from 'components/Link';
 import { toHtmlId, truncate } from 'shared/utils/string';
-import { alphaNumericSorter } from 'utils/sort';
+import { alphaNumericSorter } from 'shared/utils/sort';
 
 import css from './TagList.module.scss';
