+++ webui/react/src/pages/Cluster/ClusterOverview.table.tsx
@@ -2,7 +2,7 @@ import { ColumnType } from 'antd/es/table';
 
 import { V1ResourcePoolTypeToLabel } from 'constants/states';
 import { ResourcePool } from 'types';
-import { alphaNumericSorter, numericSorter } from 'utils/sort';
+import { alphaNumericSorter, numericSorter } from 'shared/utils/sort';
 
 export const columns: ColumnType<ResourcePool>[] = [
   {
