+++ webui/react/src/pages/TaskList.tsx
@@ -32,10 +32,11 @@ import { ShirtSize } from 'themes';
 import { ExperimentAction as Action, CommandState, CommandTask, CommandType } from 'types';
 import handleError from 'utils/error';
 import {
-  alphaNumericSorter, commandStateSorter, dateTimeStringSorter, numericSorter,
-} from 'utils/sort';
+  alphaNumericSorter, dateTimeStringSorter, numericSorter,
+} from 'shared/utils/sort';
 import { filterTasks, isTaskKillable, taskFromCommandTask } from 'utils/task';
 import { getDisplayName } from 'utils/user';
+import { commandStateSorter } from '../utils/sort';
 
 import css from './TaskList.module.scss';
 import settingsConfig, { DEFAULT_COLUMN_WIDTHS, Settings } from './TaskList.settings';
