+++ webui/react/src/utils/chart.ts
@@ -2,7 +2,7 @@ import { Theme } from 'shared/themes';
 import { Primitive, Range } from 'shared/types';
 import { ColorScale } from 'shared/utils/color';
 import { clone } from 'shared/utils/data';
-import { primitiveSorter } from 'utils/sort';
+import { primitiveSorter } from 'shared/utils/sort';
 
 /* Color Scales */
