+++ webui/react/src/utils/metric.ts
@@ -1,6 +1,6 @@
 import { RecordKey } from 'shared/types';
 import { MetricName, MetricType, WorkloadGroup } from 'types';
-import { metricNameSorter } from 'utils/sort';
+import { metricNameSorter } from './sort';
 
 export const extractMetricNames = (workloads: WorkloadGroup[]): MetricName[] => {
   const trainingNames = workloads
